requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2
django-admin-interface==0.30.1
django-auditlog==3.2.1
django-colorfield==0.14.0
django-cors-headers==4.3.1
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
phonenumbers==9.0.17
pillow==12.0.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
uritemplate==4.2.0
django-ratelimit==3.0.1
pytz==2025.2
